pyproject.toml
[tool.poetry]
name = "autocontract"
version = "0.1.0"
description = ""
authors = ["P3rdigas <pr.perdigao@campus.fct.unl.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.0"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
customtkinter = "^5.2.2"
pillow = "^10.2.0"
ctkmenubar = {git = "https://github.com/P3rdigas/CTkMenuBar.git", rev = "poetry"}
ctktooltip = {git = "https://github.com/P3rdigas/CTkToolTip-Poetry.git", rev = "poetry"}
ctkxyframe = {git = "https://github.com/P3rdigas/CTkXYFrame.git", rev = "poetry"}
docx2pdf = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
